+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 1.0a25 (unreleased)
 -------------------
 
+- Add fullobjects parameter to content GET request.
+  [timo]
+
+- Remove "sharing" attributes from GET response.
+  [timo,jaroel]
+
 - Add uninstall profile [davilima6]